Check that last line is empty before removing
This is not consistent across formatter orz
Eg. black adds a newline after the output. isort does not.

Best thing I can come up with now is check that it is an empty line
before we remove it. But this still means if that empty line was meant
to be there, and the formatter does not add a new line, we remove it
when we shouldn't