Use bang as force write (#3)
* Don't overwrite unsaved changes unless bang is used

* Resolve all formatters before starting

- In case the functions need to resolve buffer related things, they
should all get resolved synchronously

* Skip over formatter that exit with error code

- We don't want to overwrite the buffer with nothing or error output.

* Don't modify the buffer if we don't have to

Checks if input and output are different before writing to the buffer

* FormatWrite and user auto commands

- adds `FormatWrite` command to write the format changes
- adds `FormatterPost` and `FormatterPre` autocommands

* Remove last newline from output

that newline is just the end of stdout or stderr. It should not be
written to the buffer

* Check that last line is empty before removing

This is not consistent across formatter orz
Eg. black adds a newline after the output. isort does not.

Best thing I can come up with now is check that it is an empty line
before we remove it. But this still means if that empty line was meant
to be there, and the formatter does not add a new line, we remove it
when we shouldn't