Use sed as a pipe (the default behaviour)
In-place mode varies between GNU and BSD (incl. macOS) sed: GNU accepts
e.g. -i.bak or -i; BSD wants -i .bak or -i ''. It's therefore difficult
to invoke this in a cross-platform way as far as I know; you'd need to
detect which one is being used or require that GNU sed is installed.

However, since the default behaviour of sed is to read from stdin and
write to stdout, and formatter.nvim supports this behaviour quite
straightforwardly, we can just do that instead.